+++ src/assets/scss/main.scss
@@ -70,7 +70,6 @@ body {
 	flex-grow: 1;
 	text-align: center;
 	height:100%;
-	opacity: 0.7;
 	position: relative;
 	overflow: hidden;
 	background-image: none;
@@ -85,6 +84,7 @@ iframe {
 }
 
 .movingbox {
+	opacity: 0.7;
 	background-color: rgba(0, 0, 0, 100);
 	position:relative;
 	height:100%;
@@ -115,27 +115,19 @@ h1 {
 }
 
 @keyframes entrance1 {
-	0% {
-		right:40%;
-		opacity:0;
-	}
-	100% {
-		right:0;
-		opacity:1;
-	}
+	from {right:40%;
+		opacity:0;}
+	to {right:0;
+		opacity:1;}
 }
 
 @keyframes entrance2 {
-	0% {
-		top:40%;
-		opacity:0;
-	}
-	100% {
-		top:0;
-		opacity:1;
-	}
+	from {top:40%;
+		opacity:0;}
+	to {top:0;
+		opacity:0.7;}
 }
-
+//CHARACTERS
 .selection {
 	list-style-type: none;
 	margin: 0;
@@ -147,13 +139,11 @@ h1 {
 	overflow: hidden;
 	flex: 1;
 }
-
 #bg-ch {
 	display:block;
 	border-radius: 5px;
 	z-index: 50;
 }
-
 .team-image {
 	width:100%;
 	height:10%;
@@ -167,7 +157,6 @@ h1 {
 	background-position: center;
 	z-index: 40;
 }
-
 #build-image {
 	background-image: url(image/teams/bigbird.png);
 }
@@ -196,7 +185,6 @@ h1 {
 	animation-name: team-hover-out;
 	animation-duration: 0.5s;
 }
-
 .team {
 	position: absolute;
 	display:none;
@@ -205,18 +193,17 @@ h1 {
 	width:100%;
 	height:90%;
 	margin:0 auto;
-//	opacity:0.3;
 	background-color: black;
+	border-radius: 5px;
 }
 .team-name {
 	position: relative;
 	top:60%;
-	font-size: 7vw;
+	font-size: 7vmin;
 	left:9vw;
 	padding-bottom: 1em;
 	width:10vw;
 	height:auto;
-	//left:0;
 	transform: rotate(-90deg);
 	opacity:0.5;
 	z-index:60;
@@ -247,10 +234,9 @@ h1 {
 }
 .strip {
 width: 1000%;
-height:50%;
-margin-top:15%;
+height:60%;
+margin-top:5%;
 margin-left:100%;
-//background-color: red;
 -webkit-animation-name: move;
 -webkit-animation-duration: 4s;
 -webkit-animation-iteration-count: infinite;
@@ -260,12 +246,10 @@ animation-duration: 10s;
 animation-iteration-count: infinite;
 animation-timing-function: linear;
 }
-
 .strip:hover {
 -webkit-animation-play-state: paused;
 animation-play-state: paused;
 }
-
 @-webkit-keyframes move {
 from {margin-left: 100%;}
 to {margin-left: -480%;}
@@ -274,16 +258,43 @@ to {margin-left: -480%;}
 from {margin-left: 100%;}
 to {margin-left: -480%;}
 }
-
 .person {
 	background: grey;
-	color: #000;
-	width:4%;
+	width:5%;
 	height:100%;
 	margin-right:10%;
-
-		display: inline-block;
-		position: relative;
-		right:0;
-		z-index: 95;
+	display: inline-block;
+	position: relative;
+    right:0;
+	z-index: 95;
+	color: #fff;
+	text-align: center;
+	padding:1em;
+	overflow-y: auto;
+}
+#hovertext {
+	text-align: center;
+	position: relative;
+	margin:auto;
+	padding-top: 5%;
+}
+//MEMBERS
+.personimage {
+	float:right;
+	width:50%;
+	height:50%;
+	background-size: contain;
+	background-repeat: no-repeat;
+	background-position: center;
+}
+#person-1 {
+	background-image: url(image/members/1.jpg);
+}
+.persontext {
+	text-align: justify;
+	display: block;
+	font-size: 2vmin;
+}
+h3 {
+	text-align : left;
 }
