--- src/pages/characters.en.js
@@ -1,52 +1,0 @@
-import React from 'react'
-import { Link } from 'gatsby'
-const SecondPage = () => (
-  <div class="movingbox" id="characters-box">
-    <ul class="selection">
-    <div class="team" id="bg-ch"></div>
-      <li class="team-options" id="build-option">
-        <img class="team-image" id="build-image" src="data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA"></img>
-      <p class="team-name">BUILD</p>
-      <div class="team" id="build">
-        <div class="strip">
-        <div class="person">aha2323haha</div>
-        <div class="person">hoh2323oho</div>
-        <div class="person">hoh2323oho</div>
-        <div class="person">hoh2323oho</div>
-        </div>
-      </div>
-      </li>
-      <li class="team-options" id="kiosk-option">
-      <img class="team-image" id="kiosk-image" src="data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA"></img>
-      <p class="team-name">KIOSK</p>
-      <div class="team" id="kiosk"></div>
-      </li>
-      <li class="team-options" id="programming-option">
-      <img class="team-image" id="programming-image" src="data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA"></img>
-      <p class="team-name">PROGRAMMING</p>
-      <div class="team" id="programming"></div>
-      </li>
-
-      <li class="team-options" id="journalism-option">
-      <img class="team-image" id="journalism-image" src="data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA"></img>
-      <p class="team-name">JOURNALISM</p>
-     <div class="team" id="journalism"></div>
-      </li>
-
-      <li class="team-options" id="video-option">
-      <img class="team-image" id="video-image" src="data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA"></img>
-      <p class="team-name">VIDEO</p>
-      <div class="team" id="video"></div>
-      </li>
-
-      <li class="team-options" id="web-option">
-      <img class="team-image" id="web-image" src="data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA"></img>
-      <p class="team-name">WEB</p>
-      <div class="team" id="web"></div>
-      </li>
-    </ul>
-
-  </div>
-)
-
-export default SecondPage
