+++ src/pages/en/members/example.js
@@ -1,0 +1,10 @@
+import React from 'react'
+const Example = () => (
+<div>
+<h3>Person 1</h3>
+<img class="personimage" id="person-1" src="data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA"></img>
+<div id="left"><p class="persontext">Sit nihil sapiente id. Deserunt unde laborum sed eveniet et non non. Est pariatur est vel dolorem. Aspernatur rerum autem corporis excepturi esse deleniti.</p></div>
+
+</div>
+)
+export default Example
